requirements.txt
torch
torchvision
gradio       # or streamlit, depending on your choice
ultralytics
Pillow
albumentations
